Ultima Actualizacion 19/08/2024

+++ utils/save_file.py
@@ -245,14 +245,26 @@ def generate_summary_for_constraint(base_dir, output_path, constraint):
     final_df.to_csv(output_path, index=False)
     
 def save_results_to_json(data, problem_name, directory, problem_prefix, bchm):
-    directory = f"{directory}/json"
+    directory = f"{directory}/_json"
     ensure_directory_exists(directory)
     filename = f"{directory}/{problem_prefix}_violations_{problem_name}.json"
-    with open(filename, 'w') as json_file:
-        json.dump(data, json_file, indent=4)
+    if os.path.exists(filename):
+        with open(filename, 'r+') as json_file:
+            try:
+                existing_data = json.load(json_file)
+            except json.JSONDecodeError:
+                existing_data = {}
+            json_file.seek(0)
+            if not isinstance(existing_data, dict):
+                existing_data = {}
+            existing_data.update(data)
+            json.dump(existing_data, json_file, indent=4)
+    else:
+        with open(filename, 'w') as json_file:
+            json.dump(data, json_file, indent=4)
         
 def load_results_from_json(problem_name, directory, problem_prefix):
-    directory = f"{directory}/json"
+    directory = f"{directory}/_json"
     filename = f"{directory}/{problem_prefix}_violations_{problem_name}.json"
     
     if os.path.exists(filename):
